Flink update 1.16.2, added Rocksdb metrics, remove some of the global checkpoint settings to leave it to local job configuration
- Upgraded from Flink 1.16.1 to 1.16.2 testGateway
- Rocksdb metrics are inserted into flink-config to allow better debugging and optimization
- Removed local Checkpoint configurations. This will be done by the jobs in future
- Add debug mode to to Gateway: Do not delete the deployment directories in /tmp when DEBUG is set to 'true'

+++ FlinkSqlGateway/gateway.js
@@ -24,6 +24,7 @@ const bodyParser = require('body-parser');
 
 const logger = require('./lib/logger.js');
 const port = process.env.SIMPLE_FLINK_SQL_GATEWAY_PORT || 9000;
+const debug = process.env.DEBUG || 'false';
 const flinksubmit = '/opt/flink/bin/flink run';
 const runningAsMain = require.main === module;
 
@@ -102,16 +103,27 @@ function apppost (request, response) {
 
     const command = createCommand(dirname);
     submitJob(command, response).then(
-      () => { fs.rmSync(dirname, { recursive: true, force: true }); }
+      () => {
+        if (debug === 'false') {
+          logger.info('Removing ' + dirname + ': debug = ' + debug);
+          fs.rmSync(dirname, { recursive: true, force: true });
+        }
+      }
     ).catch(
       (e) => {
         logger.error(e.stack || e);
-        fs.rmSync(dirname, { recursive: true, force: true });
+        if (debug === 'false') {
+          logger.info('Removing ' + dirname + ': debug = ' + debug);
+          fs.rmSync(dirname, { recursive: true, force: true });
+        }
       }
     );
   } catch (e) {
     logger.error('Could not submit job: ' + e.stack || e);
-    fs.rmSync(dirname, { recursive: true, force: true });
+    if (debug === 'false') {
+      logger.info('Removing ' + dirname + ': debug = ' + debug);
+      fs.rmSync(dirname, { recursive: true, force: true });
+    }
   }
 }
